Added code for *Send and *Broadcast methods. Working on fleshing out delta logic; added n_state module. Transitioning m_delta module from prototype to real deltas implementation using buffers.

+++ src/n_net.c
@@ -367,7 +367,7 @@ void N_PrintAddress(FILE *fp, int peernum) {
   );
 }
 
-void N_DisconnectPlayer(int playernum) {
+void N_DisconnectPlayer(short playernum) {
   int peernum = N_GetPeerNumForPlayer(playernum);
   netpeer_t *np = NULL;
